Implement cron based labeler
Use actions/stale as an example of how to implement pagination and
rate limiting support and adapt it to the actions/labeler repository.

The cron based labeler sorts the pull request by date updated in
descending order and processes them subject to the rate limit. The
idea is that if the cron job runs often enough it will process all
the pull requests since the most recently updated ones are processed
first.